// src/types/order-buy-detail.d.ts
// 購入履歴詳細
export type OrderBuyDetail = {
  // 注文番号
  odrId: string;
  // 書式整形済注文番号
  formatedOdrNum: string;
  // 注文日時
  odrDttm: string;
  // 注文状況コード
  deliveryOdrStatusId: string;
  // 合計金額
  totalPaymentIncTax: number;
  // 合計額の内税額
  taxAmount: number;
  // 支払い方法名
  paymentStatusName: string;
  // 支払い方法コード
  paymentStatusId: number;
  // 受取方法
  rcvMethodId: number;
  // 受取方名
  rcvMethodName: string;
  // 出荷日
  shippedDt: string;
  // 送状番号
  delivTrackingId: string;
  // 商品金額
  productTotalPrice: number;
  // まとめ買い値引き額
  isBundlePrice: number;
  // ついで買い値引き額
  isSetDiscountPrice: number;
  // クーポンによる値引き額
  discountedValueByCoupon: number;
  // 送料
  shippingCharge: number;
  // 利用ポイント数
  usedPoint: number;
  // 付与ポイント数
  gainedPoint: number;
  // トクトク交換査定額
  tokutokuPrice: number;
  // お受取額
  recievePrice: number;
  // 買取上限額
  maxPurchasePrice: number;
  // WEB領収書URL
  webReceiptURL: string;
  // 保証のLINE連携ステータス
  lineTransferStatus: number;
  // お届け情報
  deliveryInfo: Array<{
    // 受取方法
    rcvMethodId: number;
    // お届け希望日
    preferedDelivDt?: string;
    // お届け希望時間帯
    deliveryTimeZone?: string;
    // 宅配ボックス
    delivBoxUseId?: 1 | 2;
    // お届け先氏名（姓）
    delivLastName?: string;
    // お届け先氏名（名）
    delivFirstName?: string;
    // お届け先氏名カナ（姓）
    delivLastNameKana?: string;
    // お届け先氏名カナ（名）
    delivFirstNameKana?: string;
    // お届け先郵便番号
    delivZipCode?: string;
    // お届け先住所１
    delivAddress1?: string;
    // お届け先住所２
    delivAddress2?: string;
    // お届け先住所３
    delivAddress3?: string;
    // 連絡先電話番号
    delivPhone?: string;
  }>;
  // 受取店舗情報
  rcvShopInfo?: Array<{
    // 受取店舗コード
    shopCode: string;
    // 受取店舗名
    shopName: string;
    // 受取店舗電話番号
    shopTel: string;
    // 受取店舗住所
    shopAddr: string;
    // お客様情報×１
    customerInfo: Array<{
      // 氏名（姓）
      lastName: string;
      // 氏名（名）
      firstName: string;
      // 氏名カナ（姓）
      lastNameKana: string;
      // 氏名カナ（名）
      firstNameKana: string;
      // 郵便番号
      zipCode: string;
      // 住所１
      address1: string;
      // 住所２
      address2: string;
      // 住所３
      address3: string;
      // 連絡先電話番号
      phone: string;
      // メールアドレス
      email: string;
    }>;
  }>;
  // 明細情報
  detailInfo: Array<OrderBuyProductDetail>;
};

// 注文商品
export type OrderBuyProductDetail = {
  // 行番号
  rowNo: number;
  // 新品中古区分
  isChuko: 1 | 2;
  // 商品コード
  janCode: string;
  // 画像パス
  imagePath: string;
  // 商品名
  itemName: string;
  // 単価（税込）
  unitPrice: number;
  // 数量
  count: number;
  // なんでも下取り価格
  tradeInPrice: number;
  // まとめ買い割引
  isBundlePrice: number;
  // 延長保証料
  extWarrantyCharge: number;
  // 設置・配線対応料金
  fieldSvcFee: number;
  // リサイクル料金
  recycleFee: number;
  // ラッピング金額
  wrappingPrice: number;
  // クーポン利用
  discountedValueByCoupon: number;
  // 小計
  subtotal: number;
  // 予約番号
  reserveNo?: number;
  // 納期表示用商品ステータス文言
  deliveryPrdStatusNote: string;
};